don't calculate unneeded LD
In simulated_z_null and simulated_z_score, LD was calculated for all SNPs in freq even if user passed the "snps" argument to specify a subset. This change saves a lot of memory if you have a whole-genome freq df, but only want to simulate smaller regions at a time.